requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
numpy
pandas
pydantic
time
json